+++ pkg/apis/backup/v1/backup_policy.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -72,6 +72,10 @@ func (a *ArangoBackupPolicy) NewBackup(d *api.ArangoDeployment) *ArangoBackup {
 
 			Labels:      d.Labels,
 			Annotations: d.Annotations,
+
+			OwnerReferences: []meta.OwnerReference{
+				d.AsOwner(),
+			},
 
 			Finalizers: []string{
 				FinalizerArangoBackup,